Godeps/_workspace/src/github.com/expanse-project/ethash/test/python/requirements.txt
pyethereum==0.7.522
nose==1.3.4
pysha3==0.3